feat(error-tracking): show the support report widget on the issue page
Replaces the header chip, which read as a stray purple pill in a row of plain
controls. Error tracking now renders the same entry the support ticket thread
renders, so a self-driving investigation looks identical wherever someone
meets it.

The entry moves into lib/signals and both products use it. Error tracking drops
the ticket thread's private-note lock, which has no meaning without a
customer-facing thread, and names errors rather than tickets in the tooltip.

It sits between the issue header and the two columns. That zone is issue-scoped,
where the right column is scoped to the selected occurrence. Its width is capped
so the title and the fix state stay as close as they are in the thread.
