+++ products/error_tracking/frontend/scenes/ErrorTrackingIssueScene/ErrorTrackingIssueScene.tsx
@@ -57,7 +57,7 @@ import {
     errorTrackingIssueSceneLogic,
 } from './errorTrackingIssueSceneLogic'
 import { IssueEventsPanel } from './IssueEventsPanel'
-import { LinkedReportsIndicator } from './LinkedReportsIndicator'
+import { LinkedReports } from './LinkedReports'
 import { ErrorTrackingIssueScenePanel } from './ScenePanel'
 import { IssueAssigneeSelect } from './ScenePanel/IssueAssigneeSelect'
 
@@ -165,7 +165,6 @@ export function ErrorTrackingIssueScene(): JSX.Element {
                                     actions={
                                         isMobile ? undefined : (
                                             <div className="flex items-center gap-1">
-                                                <LinkedReportsIndicator />
                                                 <StatusIndicator status={issue.status} withTooltip />
                                                 <IssueAssigneeSelect
                                                     assignee={issue.assignee}
@@ -204,7 +203,6 @@ export function ErrorTrackingIssueScene(): JSX.Element {
 
                                 {isMobile && (
                                     <div className="flex items-center gap-1.5 px-2 py-1.5 border-b flex-wrap">
-                                        <LinkedReportsIndicator />
                                         <StatusIndicator status={issue.status} withTooltip />
                                         <IssueAssigneeSelect
                                             assignee={issue.assignee}
@@ -225,6 +223,8 @@ export function ErrorTrackingIssueScene(): JSX.Element {
                                 )}
 
                                 <ErrorTrackingIssueScenePanel issue={issue} />
+
+                                <LinkedReports />
 
                                 <div className="ErrorTrackingIssue flex flex-grow min-h-0 overflow-hidden">
                                     <div className="relative flex flex-1 h-full w-full min-h-0">
